make lrv3 work

+++ dmlib/src/main/java/nyh/classify/lr/v3/Instance.java
@@ -1,10 +1,10 @@
 package nyh.classify.lr.v3;
 
 public class Instance {
-    public int label;
+    public double label;
     public double[] x;
     public Instance(){}
-    public Instance(int label,double[] x){
+    public Instance(double label,double[] x){
         this.label = label;
         this.x = x;
     }
